requirements.txt
pillow>=9.0.0
numpy>=1.20.0
pyyaml>=6.0
